fix: 이력 조회 시 타입 에러로 인한 수정, feat: AWS에 올릴 도커 및 컴포즈 파일 생성

+++ src/main/kotlin/com/project/ecoNrich/domain/employee/rqrs/EmployeeHistoryRs.kt
@@ -17,7 +17,7 @@ data class EmployeeHistoryRs(
     val jobTitle: String? = null,
 
     @Schema(description = "부서 ID")
-    val departmentId: Int? = null,
+    val departmentId: Long? = null,
 
     @Schema(description = "부서 이름")
     val departmentName: String? = null,
